+++ src/components/BalanceIncome.js
@@ -5,28 +5,25 @@ const Container = styled.div`
 display: flex;
 flex-direction: column;
 align-items: center;
-width: 100%;
-gap: 0.25rem;
+width: 100vw;
+border-right: 1px solid black;
 
-
-.negative {
-    color: ${({ theme }) => theme.colors.negative};
-    font-weight: 600;
+.positive {
+    color: green;
 }
 `
 
-
-export const BalanceExpense = ({ transactions }) => {
+export const BalanceIncome = ({ transactions }) => {
     const amounts = transactions.map(transaction => transaction.amount);
 
-    const expense = amounts
-        .filter(amount => amount < 0)
+    const income = amounts
+        .filter(amounts => amounts > 0)
         .reduce((acc, amount) => (acc += amount), 0).toFixed(2);
 
     return (
         <Container>
-            <h3>Expense</h3>
-            <p className={`${expense < 0 ? 'negative' : null}`}>${expense}</p>
+            <h3>Income</h3>
+            <p className={`${income > 0 ? 'positive' : null}`}>${income}</p>
         </Container>
     )
 }
