+++ src/components/TransactionItemText.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const TransactionItemText = ({ text }) => {
+    return (
+        <>
+            <h3>{text}</h3>
+        </>
+    )
+}
