--- src/context/app-actions.js
@@ -1,5 +1,0 @@
-export const ADD_TRANSACTION = 'ADD_TRANSACTION';
-export const DELETE_TRANSACTION = 'DELETE_TRANSACTION';
-export const TOGGLE_TRANSACTION = 'TOGGLE_TRANSACTION';
-export const FILTER_YEAR = 'FILTER_YEAR';
-
