--- src/context/app-context.js
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const AppContext = createContext();
-
-export default AppContext;
