+++ src/index.css
@@ -1,5 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap');
-
+@import-normalize;
 
 *,
 *::before,
@@ -15,7 +15,4 @@ body {
   justify-content: center;
   align-items: center;
   height: auto;
-  background-color: #d4d4d4;
-  color: hsl(0, 0%, 25%);
-
 }
